pyproject.toml
[tool.poetry]
description = "Water level anomaly detection"
authors = ["Sajad Safarveisi"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.10"
requests = "^2.32.3"
pandas = "^2.2.2"
matplotlib = "^3.9.0"
scikit-learn = "^1.6.0"
streamlit = "^1.41.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"
ipywidgets = "^8.1.3"
black = {extras = ["jupyter"], version = "^24.10.0"}
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"